workshops/FDA_Doc_Search/requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.13
tqdm==4.66.5
lxml==5.3.0
gradio==4.43.0
python-dotenv==1.0.1
